requirements.txt
streamlit
streamlit-aggrid
fastapi
uvicorn
sqlalchemy
plotly
unidecode
google-api-python-client
google-auth-httplib2
google-auth-oauthlib